requirements.txt
cx-oracle==5.3
decorator==4.0.11
Django==1.10.6
django-bootstrap3==8.2.3
django-celery==3.2.1
django-tables2==1.5.0
django-windows-tools==0.2
djangorestframework
Jinja2==2.9.4
